refactor(handlers): extract dispatch.ts; intent.ts is now a 39-line barrel (Stage B)
handleToolCall (the dispatch pipeline) + the LLM error-formatting tree +
TOOL_SCOPES/hasRequiredScope move to dispatch.ts. intent.ts is reduced to a
deprecated back-compat barrel re-exporting the public surface. Completes the
8,199 → 39 line breakup of intent.ts. 3672 tests, A1 zero-diff, barrel-surface
locked.
